modification following the recommendation of Mr. Catoire
